feat: Add menu item modifiers and detail modal
Introduces the ability to add modifiers to menu items, allowing for customization such as size, toppings, or add-ons. A new item detail modal has been implemented in the Point of Sale component to facilitate the selection of these modifiers and capture custom notes before adding an item to the order. The Kitchen Order Card has also been updated to display selected modifiers.